% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{package}
\name{csbELISA}
\alias{csbELISA-package}
\alias{csbELISA}
\title{Analyse and visualize ELISA data}
\description{
The csbELISA R package is an open-source project that provides a
user-friendly, interactive UI for automated ELISA data processing,
standard curve fitting, concentration estimation, visualization,
and statistical analysis.
}
\author{
\strong{Maintainer}: Caio Santos Bonilha \email{Caio.Bonilha@glasgow.ac.uk}

}
