% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runCsbELISA}
\alias{runCsbELISA}
\title{Run the csbELISA application}
\usage{
runCsbELISA()
}
\description{
runCsbELISA() launches the csbELISA Shiny application.
}
\details{
Use runCsbELISA() at the R console to initiate the csbELISA UI.
}
